terminando ACL

+++ app/Http/Controllers/Admin/UsersController.php
@@ -13,29 +13,39 @@ class UsersController extends Controller
 
     public function index()
     {
+        $this->authorize('user_list');
+        
         $users = User::all();
         return view('admin.users.index', compact('users'));
     }
 
     public function create()
     {
+        $this->authorize('user_add');
+        
         return view('admin.users.create');
     }
 
     public function store(Request $request)
     {
+        $this->authorize('user_add');
+        
         User::create($this->prepareFields($request));
         return redirect()->route('admin.users.index');
     }
 
     public function edit($id)
     {
+        $this->authorize('user_edit');
+        
         $user = User::find($id);
         return view('admin.users.edit', compact('user'));
     }
 
     public function update(Request $request, $id)
     {
+        $this->authorize('user_edit');
+        
         User::find($id)->update($this->prepareFields($request));
         return redirect()->route('admin.users.index');
     }
@@ -43,12 +53,16 @@ class UsersController extends Controller
 
     public function destroy($id)
     {
+        $this->authorize('user_destroy');
+        
         User::find($id)->delete();
         return redirect()->route('admin.users.index');
     }
     
     public function roles($id)
     {
+        $this->authorize('user_view_roles');
+        
         $user = User::find($id);
         $roles = Role::all();
         return view('admin.users.roles', compact('user', 'roles'));
@@ -56,6 +70,8 @@ class UsersController extends Controller
 
     public function storeRole(Request $request, $id)
     {
+        $this->authorize('user_add_role');
+        
         $user = User::find($id);
         $roles = Role::findOrFail($request->all()['role_id']);
         
@@ -65,6 +81,8 @@ class UsersController extends Controller
     
     public function revokeRole($id, $role_id)
     {
+        $this->authorize('user_revoke_role');
+        
         $user = User::find($id);
         $roles = Role::findOrFail($role_id);
